requirements.txt
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
packaging
python-telegram-bot
langchain_core
langchain_google_genai